Refactor traffic sign handling in CMake and update data structures for TsrHeader

+++ apps/rpi5-firmware/Utils.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 
 struct FrameHeader {
 	uint32_t frameNbr;
@@ -8,13 +9,13 @@ struct FrameHeader {
 };
 
 struct TsrHeader {
-    uint32_t frameNbr;
-    uint16_t numDetections;
-    int      trafficSign;
-    float    accuracy;
-    int      x;
-    int      y;
-    int      width;
-    int      height;
+    uint32_t	frameNbr;
+    uint16_t	numDetections;
+    uint16_t    trafficSign;
+    float		accuracy;
+    uint32_t    x;
+    uint32_t    y;
+    uint32_t    width;
+    uint32_t    height;
 };
